+++ middleware/std/std_test.go
@@ -85,3 +85,74 @@ func TestMiddleware(t *testing.T) {
 		})
 	}
 }
+
+func TestProvider(t *testing.T) {
+	tests := map[string]struct {
+		handlerID   string
+		config      middleware.Config
+		req         func() *http.Request
+		mock        func(m *mmetrics.Recorder)
+		handler     func() http.Handler
+		expRespCode int
+		expRespBody string
+	}{
+		"A default HTTP middleware should call the recorder to measure.": {
+			req: func() *http.Request {
+				return httptest.NewRequest(http.MethodPost, "/test", nil)
+			},
+			mock: func(m *mmetrics.Recorder) {
+				expHTTPReqProps := metrics.HTTPReqProperties{
+					ID:      "/test",
+					Service: "",
+					Method:  "POST",
+					Code:    "202",
+				}
+				m.On("ObserveHTTPRequestDuration", mock.Anything, expHTTPReqProps, mock.Anything).Once()
+				m.On("ObserveHTTPResponseSize", mock.Anything, expHTTPReqProps, int64(15)).Once()
+
+				expHTTPProps := metrics.HTTPProperties{
+					ID:      "/test",
+					Service: "",
+				}
+				m.On("AddInflightRequests", mock.Anything, expHTTPProps, 1).Once()
+				m.On("AddInflightRequests", mock.Anything, expHTTPProps, -1).Once()
+			},
+			handler: func() http.Handler {
+				return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+					w.WriteHeader(202)
+					w.Write([]byte("Я бэтмен")) // nolint: errcheck
+				})
+			},
+			expRespCode: 202,
+			expRespBody: "Я бэтмен",
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			assert := assert.New(t)
+			require := require.New(t)
+
+			// Mocks.
+			mr := &mmetrics.Recorder{}
+			test.mock(mr)
+
+			// Create our negroni instance with the middleware.
+			test.config.Recorder = mr
+			m := middleware.New(test.config)
+			provider := stdmiddleware.HandlerProvider(test.handlerID, m)
+			h := provider(test.handler())
+
+			// Make the request.
+			resp := httptest.NewRecorder()
+			h.ServeHTTP(resp, test.req())
+
+			// Check.
+			mr.AssertExpectations(t)
+			assert.Equal(test.expRespCode, resp.Result().StatusCode)
+			gotBody, err := ioutil.ReadAll(resp.Result().Body)
+			require.NoError(err)
+			assert.Equal(test.expRespBody, string(gotBody))
+		})
+	}
+}